Moved logger file to singleton Move: - Move logger to singleton.
