+++ internal/services/iptable/ipTableRule.go
@@ -6,7 +6,7 @@ import (
 	"os/exec"
 	"strings"
 
-	"github.com/Seicrypto/torcontroller/internal/services/logger"
+	"github.com/Seicrypto/torcontroller/internal/singleton/logger"
 )
 
 // applyIptablesRules applies iptables rules for redirecting traffic to port 8118.
