+++ internal/singleton/logger/index.go
